+++ k8s.tf
@@ -64,3 +64,34 @@ resource "kubernetes_stateful_set" "elasticsearch" {
     }
   }
 }
+
+
+resource "kubernetes_cron_job" "backup_cleanup" {
+  metadata {
+    name = "elasticsearch-backup"
+  }
+  spec {
+    concurrency_policy            = "Replace"
+    failed_jobs_history_limit     = 5
+    schedule                      = "0 3 * * *"
+    successful_jobs_history_limit = 3
+    job_template {
+      metadata {}
+      spec {
+        backoff_limit              = 2
+        ttl_seconds_after_finished = 10
+        template {
+          metadata {}
+          spec {
+            container {
+              name    = "elasticsearch-backup-cleanup"
+              image   = "curlimages/curl"
+              command = ["/bin/sh", "-c", "curl -s -XPOST http://es-ilb.${google_compute_forwarding_rule.elasticsearch.name}.il4.${var.region}.lb.${var.project}.internal:9200/_snapshot/${local.backup_repository}/_cleanup?pretty"]
+            }
+            restart_policy = "OnFailure"
+          }
+        }
+      }
+    }
+  }
+}